fix big about camera can not use

+++ app/src/main/java/com/xiangchuang/risks/view/CompanyActivity.java
@@ -138,7 +138,7 @@ public class CompanyActivity extends BaseActivity {
         pop.setFocusable(true);
         pop.setOutsideTouchable(true);
         pop.setContentView(popview);
-        
+
         searchEdit.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
@@ -383,15 +383,20 @@ public class CompanyActivity extends BaseActivity {
 
     @Override
     public void onBackPressed() {
-        long secondTime = System.currentTimeMillis();
-        if (secondTime - firstTime > 2000) {
-            Toast.makeText(this, "再按一次退出程序", Toast.LENGTH_SHORT).show();
-            firstTime = secondTime;
-        } else {
+        if ("com.xiangchuangtec.luolu.animalcounter".equals(AppConfig.getAppContext().getPackageName())) {
+            long secondTime = System.currentTimeMillis();
+            if (secondTime - firstTime > 2000) {
+                Toast.makeText(this, "再按一次退出程序", Toast.LENGTH_SHORT).show();
+                firstTime = secondTime;
+            } else {
 //            finish();
-            android.os.Process.killProcess(android.os.Process.myPid());
-            System.exit(0);
+                android.os.Process.killProcess(android.os.Process.myPid());
+                System.exit(0);
+            }
+        } else {
+            super.onBackPressed();
         }
+
     }
 
 }
